==> 3_Relation Extraction and Synset Grouping/1_prepare_synset_data.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd drive/MyDrive/Colab\ Notebooks/Synset_Grouping

ls

Created files: 
- 'vp_list.csv' : VP, Main Verb, Lemmatized Verb
- 'lemma_list.csv' : Lemmatized Verb, Frequency, Synset, Status

import time

start_time = time.time()

import spacy   
import numpy as np
import pandas as pd
from spacy.matcher import Matcher                                                                                                                                                                                         
nlp = spacy.load('en_core_web_sm')  # download model first
from nltk import tokenize

from itertools import chain
from nltk.corpus import wordnet
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer

### Load dataset and tokenize

data = pd.read_csv('preprocessed_topic4_corpus.csv')
data.head()

data.drop(data[data.isna().any(axis=1)].index, axis=0, inplace=True)

len(data)

data.reset_index(inplace = True)

import nltk
nltk.download('punkt')

# import nltk
nltk.download('wordnet')
nltk.download('omw-1.4')

vp_pattern=[{'POS': 'VERB', 'OP': '?'},{'POS': 'ADV', 'OP': '*'},{'POS': 'AUX', 'OP': '*'}, {'POS': 'VERB', 'OP': '+'}]
vp_matcher = Matcher(nlp.vocab) 
vp_matcher.add("verb-phrases", [vp_pattern], greedy = "LONGEST")

main_verb_pattern = [{'POS': 'VERB', 'OP': '+'}]
main_verb_matcher = Matcher(nlp.vocab) 
main_verb_matcher.add("main-verb", [main_verb_pattern])

vp_list = []
main_verb_list = []

# df columns 

# doc_id, sent_id, word_id, word, cluster_title(grouped title/ taken verb)

sent_num = len(data)

for idx in range(sent_num):
  sentence = data.loc[idx, 'abstract_sents']
  doc = nlp(sentence) 
  
    # VP extraction 
  vp_matches = vp_matcher(doc) 
  vp_list = [str(doc[start:end]) for _, start, end in vp_matches]

  start_list = [start for _, start, end in vp_matches]

  if vp_list:
    data.at[idx,'verb'] = np.array(vp_list, dtype=object)
    data.at[idx,'verb_id'] = np.array(start_list, dtype=object)
  else:
    data.at[idx,'verb'] = np.nan
    data.at[idx,'verb_id'] = np.nan

data.dropna(axis=0, inplace=True)

data.head()

data = data.explode(['verb', 'verb_id'])

topic4_verb_df = data[['doc_id', 'sent_id', 'verb_id', 'verb']]

topic4_verb_df.head()

# lemmatized verb

topic4_verb_df['check_verb'] = topic4_verb_df['verb'].apply(lambda x: main_verb_matcher(nlp(x)) if main_verb_matcher(nlp(x)) else np.nan)

topic4_verb_df.head()

topic4_verb_df.dropna(axis=0, inplace=True)

topic4_verb_df['main_verb'] = topic4_verb_df['verb'].apply(lambda x: [str(nlp(x)[start:end]) for _, start, end in main_verb_matcher(nlp(x))][-1])

topic4_verb_df['lemma_verb'] = topic4_verb_df['main_verb'].apply(lambda x: WordNetLemmatizer().lemmatize(x, pos ="v"))

topic4_verb_df.head()

topic4_verb_df.to_csv('preprocessed_topic4_verb_list.csv')

end_time = time.time()

print('Estimate runtime (total): ', (end_time-start_time)/60, ' minutes.')

# for 5000 sents
# Estimate runtime (total):  5.037542378902435  minutes.

# for all
#Estimate runtime (total):  42.84870324532191  minutes.
