==> 3_Relation Extraction and Synset Grouping/2_prepare_synset_data_part2.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd drive/MyDrive/Colab\ Notebooks/Synset_Grouping

ls

Created files: 
- 'vp_list.csv' : VP, Main Verb, Lemmatized Verb
- 'lemma_list.csv' : Lemmatized Verb, Frequency, Synset, Status

import time

start_time = time.time()

import spacy   
import numpy as np
import pandas as pd
from spacy.matcher import Matcher                                                                                                                                                                                         
nlp = spacy.load('en_core_web_sm')  # download model first
from nltk import tokenize

from itertools import chain
from nltk.corpus import wordnet
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer

import nltk
nltk.download('wordnet')
nltk.download('omw-1.4')

### Load dataset and tokenize

vp_df = pd.read_csv('preprocessed_topic4_verb_list.csv')
vp_df.head()

# unique vp list with frequency and status
# status --> if there is no synsets for a verb, this might be a wrong verb (e.g. amd)-related), 
#           so if a verb doesn't have synsets, status = invalid, otherwise, status = valid

vp_freq_df = pd.DataFrame()
vp_freq_df = pd.Series(vp_df['lemma_verb']).value_counts().sort_index().reset_index().reset_index(drop=True)
vp_freq_df.columns = ['Lemmatized Verb', 'Frequency']

vp_freq_df.head()

vp_synset_df = pd.DataFrame()
vp_synset_df['Lemmatized Verb'] = ""
vp_synset_df['Frequency'] = ""
vp_synset_df['Synset'] = ""
vp_synset_df['Status'] = ""

for idx in range(len(vp_freq_df)):
    
    lemma_mv = vp_freq_df['Lemmatized Verb'][idx]
    vp_synset_df.loc[idx,'Lemmatized Verb'] = lemma_mv
    vp_synset_df.loc[idx,'Frequency'] = vp_freq_df['Frequency'][idx]
    
    # get all verb synset of mv 
    mv_syn = wordnet.synsets(lemma_mv, pos='v')
    
    # get lemma set of each synset of mv - all synonyms of mv
    mv_syn_lemma = [mv_syn[i].lemma_names() for i in range(len(mv_syn))]
    mv_syn_lemma = list(chain.from_iterable(mv_syn_lemma))
    mv_syn_lemma = list(set(mv_syn_lemma))
    #print('\nmv_syn_lemma: ',mv_syn_lemma)
    
    if mv_syn_lemma:
        vp_synset_df.loc[idx,'Synset'] = np.array(mv_syn_lemma, dtype=object)
        vp_synset_df.loc[idx,'Status'] = 'valid'
    else:
        vp_synset_df.loc[idx,'Status'] = 'invalid'

vp_synset_df.head()

vp_synset_df[vp_synset_df['Status'] == 'invalid']

vp_synset_df.to_csv('preprocessed_topic4_vp_synset.csv')

end_time = time.time()

print('Estimate runtime (total): ', (end_time-start_time)/60, ' minutes.')

# Estimate runtime (total):  0.2910415252049764  minutes.
