==> 3_Relation Extraction and Synset Grouping/3_group_synset.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd drive/MyDrive/Colab\ Notebooks/Synset_Grouping

ls

- input: <unique,valid> lemmatized verb list with frequency and synset

- output: 
    - input file updated with cluster_id and cluster_title
    - cluster file with <cluster_id, cluster_title, verb_list, actua_data_list: [(lemma, freq)]>

import pandas as pd
import numpy as np
import time

start_time = time.time()

df = pd.read_csv('preprocessed_topic4_vp_synset.csv')
print('total data: ', len(df))
df.head()

### consider for only valid verbs

valid_df = df[df['Status'] == 'valid']
print('total data: ', len(valid_df))

valid_df.columns = ['id', 'Lemmatized Verb', 'Frequency', 'Synset', 'Status']
valid_df.head()

'''
subset = 500
valid_df = valid_df[:subset]

valid_df.tail()
'''

def str2list(data):
    remove_list = ['[', ']', '\'', '\n']
    for symbol in remove_list:
        data = data.replace(symbol, '')
    lst = data.split(' ')
    
    return lst

# initialize empty df
cluster_df = pd.DataFrame()
cluster_df['cluster_id'] = ''
cluster_df['cluster_title'] = ''
cluster_df['verb_list'] = ''
cluster_df['actual_data_list'] = ''

valid_df['cluster_id'] = ''
valid_df['cluster_title'] = ''
#valid_df['match_cluster_ids'] = ''
#valid_df['match_synset_counts'] = ''

cluster_id = 0

# looping all valid verbs
for verb_idx in valid_df['id']:
    
    # get lemma and synset of each valid verb
    lemma = valid_df['Lemmatized Verb'][verb_idx]
    lemma_synset = str2list(valid_df['Synset'][verb_idx])
    lemma_freq = valid_df['Frequency'][verb_idx]
    lemma_freq_tuple = (lemma, lemma_freq)
    #print('lemma: ', lemma)
    
    # flag for cluster found or not
    cluster_found = 0
    
    # loop all currently created clusters to search for lemma
    found_cid_list = []
    intersect_synset_count_list = []
    
    for c_idx in cluster_df['cluster_id']:
        
        # get synset list in each cluster
        cluster_synset = cluster_df[cluster_df['cluster_id'] == c_idx]['verb_list'].values[0]
        
        # count # of clusters in which lemma appeared in it's synset
        if lemma in cluster_synset:
            cluster_found += 1
            
            # keep cluster_idx and overlap synset count
            intersect_synset = list(set(lemma_synset) & set(cluster_synset))
            found_cid_list.append(c_idx)
            intersect_synset_count_list.append(len(intersect_synset))
            
            #print('lemma_synset: ', lemma_synset)
            #print('cluster_synset: ', cluster_synset)
            #print('Intersect_synset: ', intersect_synset)
            
            #print('found_cid_list: \t\t', found_cid_list)
            #print('intersect_synset_count_list: \t', intersect_synset_count_list)
            
    #print('cluster found: ', cluster_found)
    
    # if lemma is not in any of existing cluster
    if not cluster_found:
        cluster_df.loc[cluster_id,'cluster_title'] = lemma
        cluster_df.loc[cluster_id,'cluster_id'] = cluster_id
        cluster_df.loc[cluster_id,'verb_list'] = np.array(lemma_synset)
        cluster_df.loc[cluster_id,'actual_data_list'] = np.array([lemma_freq_tuple])
        cluster_id = cluster_id+1
        
        valid_df.loc[verb_idx,'cluster_id'] = cluster_id-1
        #print('assigned to cluster (not found): ', cluster_id-1)
        
    else:        
        # choose highest overlap synset count and respective cluster. Take that cluster to be merged
        max_intersect_synset_count = max(intersect_synset_count_list)
        found_cid_idx = intersect_synset_count_list.index(max_intersect_synset_count)
        best_match_cid = found_cid_list[found_cid_idx]
        
        # get cluster_synset and existing data_list of best_match_cid
        best_match_cluster_synset = cluster_df[cluster_df['cluster_id'] == best_match_cid]['verb_list'].values[0].tolist()
        existing_data_list = cluster_df[cluster_df['cluster_id'] == best_match_cid]['actual_data_list'].values[0].tolist()

        # append cluster_verb_list/ modified cluster_synset
        modified_synset = list(set(best_match_cluster_synset + lemma_synset))
        cluster_df.loc[best_match_cid,'verb_list'] = np.array(modified_synset)

        # modify actual_data_list
        modified_data_list = existing_data_list + [lemma_freq_tuple]
        cluster_df.loc[best_match_cid,'actual_data_list'] = np.array(modified_data_list)
        
        #print('chosen cid: ', best_match_cid)
        
        valid_df.loc[verb_idx,'cluster_id'] = best_match_cid
        #valid_df.loc[verb_idx,'match_cluster_ids'] = np.array(found_cid_list)
        #valid_df.loc[verb_idx,'match_synset_counts'] = np.array(intersect_synset_count_list)
        #print('assigned to cluster: ', best_match_cid)
        

### assigning cluster_title according to freq


for c_id in cluster_df['cluster_id']:
    #print('cluster_id: ', c_id)
    actual_data_list = cluster_df[cluster_df['cluster_id'] == c_id]['actual_data_list'].values[0].tolist()
    #print('actual_data_list: ', actual_data_list, 'length: ', len(actual_data_list))
    
    if len(actual_data_list) > 1:
        data_list = []
        freq_list = []
        for len_idx in range(len(actual_data_list)):
            data, freq = actual_data_list[len_idx]
            data_list.append(data)
            freq_list.append(freq)
            
        freq_list = list(map(int, freq_list))
        #print('data_list: ', data_list)
        #print('freq_list: ', freq_list)
        #print('actual_data_list: ', actual_data_list, 'length: ', len(actual_data_list))
        
        max_freq = max(freq_list)
        max_idx = freq_list.index(max_freq)
        title_data = data_list[max_idx]
        
        #print('max_freq_value: ', max_freq)
        #print('max_idx: ', max_idx)
        #print('title_data: ', title_data)
        
        cluster_df.loc[c_id,'cluster_title'] = title_data

cluster_df['member_lemma'] = cluster_df['actual_data_list'].apply(lambda x: x[:,0])

cluster_df[:20]

print("total clusters: ", len(cluster_df))

cluster_df.to_csv('preprocessed_topic4_cluster_matrix.csv')

end_time = time.time()

print('Estimate runtime: ', (end_time-start_time)/60, ' minutes.')

# Estimate runtime:  1.5818639238675436  minutes.

# all
# Estimate runtime:  5.733248972892762  minutes.
