==> 3_Relation Extraction and Synset Grouping/4_label_cluster_title_to_topic4.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd drive/MyDrive/Colab\ Notebooks/Synset_Grouping

ls

- input: <unique,valid> lemmatized verb list with frequency and synset

- output: 
    - input file updated with cluster_id and cluster_title
    - cluster file with <cluster_id, cluster_title, verb_list, actua_data_list: [(lemma, freq)]>

import pandas as pd
import time

start_time = time.time()

cluster_df = pd.read_csv('preprocessed_topic4_cluster_matrix.csv')

cluster_df.head()

vp_df = pd.read_csv('topic4_verb_list_5000sents.csv')

vp_df.head()

for vp_idx in range(len(vp_df)):
  lemma_verb = vp_df.loc[vp_idx, 'lemma_verb']
  # print('vp: ', lemma_verb)

  for cluster_idx in range(len(cluster_df)):
    member_lemma_list = cluster_df.loc[cluster_idx, 'member_lemma'].replace('\n', '').strip('\'][\'').split('\' \'')

    if lemma_verb in member_lemma_list:
      vp_df.loc[vp_idx, 'cluster_title'] = cluster_df.loc[cluster_idx, 'cluster_title']

      '''
      print('cluster_idx: ', cluster_idx)
      print('cluster_title: ', cluster_df.loc[cluster_idx, 'cluster_title'])
      print('cluster list: ', member_lemma_list)
      '''
  # break

vp_df.head()

vp_df

vp_df.to_csv('preprocessed_topic4_synset_group_finalResult.csv')

end_time = time.time()

print('Estimate runtime: ', (end_time-start_time)/60, ' minutes.')

# Estimate runtime:  1.5818639238675436  minutes.

# all
# Estimate runtime:  1.5682579596837363  minutes.
